pyproject.toml
[build-system]
requires = ["setuptools>=61.0"]
build-backend = "setuptools.build_meta"

[project]
name = "vibe-eval"
version = "2.0.0"
description = "Evaluate LLMs on vibe coding — building complete apps from natural language prompts"
readme = "README.md"
requires-python = ">=3.9"
license = {text = "MIT"}
authors = [
    {name = "Arjun Divecha"}
]
keywords = ["llm", "evaluation", "benchmark", "coding", "ai"]
classifiers = [
    "Development Status :: 4 - Beta",
    "Intended Audience :: Developers",
    "License :: OSI Approved :: MIT License",
    "Programming Language :: Python :: 3",
    "Programming Language :: Python :: 3.9",
    "Programming Language :: Python :: 3.10",
    "Programming Language :: Python :: 3.11",
    "Programming Language :: Python :: 3.12",
]
dependencies = [
    "anthropic>=0.30.0",
    "openai>=1.0.0",
    "google-generativeai>=0.5.0",
    "httpx>=0.25.0",
    "python-dotenv>=1.0.0",
    "rich>=13.0.0",
    "click>=8.0.0",
]

[project.scripts]
vibe-eval = "vibe_eval.cli:cli"

[project.urls]
Homepage = "https://github.com/arjundivecha/vibe-eval"
Repository = "https://github.com/arjundivecha/vibe-eval"
